backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbi-native"
version = "0.1.0"
description = "Backend for the ChatBI Native Superset extension — LangGraph agent with MCP integration"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "ChatBI Team" }]

dependencies = [
    "flask>=3.0",
    "langgraph>=0.3",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "mcp>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
